+++ source/custom-elements/drawer/input/forgot.ts
@@ -4,7 +4,7 @@ import { html, TemplateResult,  }     from 'lit';
 import { customElement, state }       from "lit/decorators.js"
 import { LazyLoader }                 from '../../../lazy-loader';
 import { sendPasswordResetEmail/*, verifyPasswordResetCode*/ }     from 'firebase/auth';
-import { auth }                       from '../../../start';
+import { auth }                       from '../../../firebase/start';
 
 @customElement('reset-email')
 export class ResetEmail extends LazyLoader {