+++ source/custom-elements/settings/input/email.ts
@@ -2,7 +2,7 @@ import { CSSResultArray, LitElement, css, html } from "lit";
 import { property, state }  from "lit/decorators.js";
 import { connect }          from "pwa-helpers";
 import store, { AppState }  from "../../../store";
-import { auth }             from "../../../start";
+import { auth }             from "../../../firebase/start";
 import { labelStyle }       from "../../../css/form/label";
 import { inputStyle }       from "../../../css/form/input";
 import { buttonStyle }      from "../../../css/form/button";