+++ source/custom-elements/settings/input/password.ts
@@ -3,7 +3,7 @@ import { CSSResultArray, LitElement, css, html } from "lit";
 import { state }            from "lit/decorators.js";
 import { connect }          from "pwa-helpers";
 import store, { AppState }  from "../../../store";
-import { auth }             from "../../../start";
+import { auth }             from "../../../firebase/start";
 import { labelStyle }       from "../../../css/form/label";
 import { inputStyle }       from "../../../css/form/input";
 import { buttonStyle  }     from "../../../css/form/button";